Create Project_Euler_2_Even_Fibonacci_numbers.cs
Each new term in the Fibonacci sequence is generated by adding the previous two terms. By starting with 1 and 2, the first  terms will be:
1, 2, 3, 5, 8, 13, 21, 34,55,89

By considering the terms in the Fibonacci sequence whose values do not exceed , find the sum of the even-valued terms.

Input Format

First line contains  that denotes the number of test cases. This is followed by  lines, each containing an integer, .

input 10 output 10
input 100 output 44