Create Project_Euler_6_Sum_square_difference.cs
The sum of the squares of the first ten natural numbers is, 1^2+2^2+...+10^ =385 . The square of the sum of the first ten natural numbers is, (1+2+..+10)^2=55^2=3025. Hence the absolute difference between the sum of the squares of the first ten natural numbers and the square of the sum is 2640 .

Find the absolute difference between the sum of the squares of the first  natural numbers and the square of the sum.

Input Format

First line contains  that denotes the number of test cases. This is followed by  lines, each containing an integer, .

Constraints